requirements.txt
requests
numpy
django-bootstrap3
django-extensions
git+git://github.com/tkrajina/gpxpy.git
git+git://github.com/thearn/pygmaps-extended.git
